Refactored some of Ast\Locator into Ast\FindReflectionsInTree
