+++ src/SourceLocator/Ast/Locator.php
@@ -4,8 +4,7 @@ namespace BetterReflection\SourceLocator\Ast;
 
 use BetterReflection\Identifier\Identifier;
 use BetterReflection\Identifier\IdentifierType;
-use BetterReflection\SourceLocator\Located\DefiniteLocatedSource;
-use BetterReflection\SourceLocator\Located\PotentiallyLocatedSource;
+use BetterReflection\SourceLocator\Ast\Strategy\NodeToReflection;
 use BetterReflection\SourceLocator\Located\LocatedSource;
 use BetterReflection\Reflection\Reflection;
 use BetterReflection\Reflector\Reflector;
@@ -17,9 +16,9 @@ use PhpParser\Lexer;
 class Locator
 {
     /**
-     * @var NodeToReflection
+     * @var FindReflectionsInTree
      */
-    private $nodeToReflection;
+    private $findReflectionsInTree;
 
     /**
      * @var Parser
@@ -28,7 +27,9 @@ class Locator
 
     public function __construct(Reflector $reflector)
     {
-        $this->nodeToReflection = new NodeToReflection($reflector);
+        $this->findReflectionsInTree = new FindReflectionsInTree(
+            new NodeToReflection($reflector)
+        );
 
         $this->parser = new Parser\Multiple([
             new Parser\Php7(new Lexer()),
@@ -80,7 +81,7 @@ class Locator
      */
     public function findReflectionsOfType(LocatedSource $locatedSource, IdentifierType $identifierType)
     {
-        return $this->reflectFromTree(
+        return $this->findReflectionsInTree->__invoke(
             $this->parser->parse($locatedSource->getSource()),
             $identifierType,
             $locatedSource
@@ -103,77 +104,5 @@ class Locator
         }
 
         throw IdentifierNotFound::fromIdentifier($identifier);
-    }
-
-    /**
-     * @param Node $node
-     * @param LocatedSource $locatedSource
-     * @param Node\Stmt\Namespace_|null $namespace
-     * @return Reflection|null
-     */
-    private function reflectNode(Node $node, LocatedSource $locatedSource, Node\Stmt\Namespace_ $namespace = null)
-    {
-        if ($locatedSource instanceof PotentiallyLocatedSource) {
-            $locatedSource = DefiniteLocatedSource::fromPotentiallyLocatedSource($locatedSource);
-        }
-
-        return $this->nodeToReflection->__invoke($node, $locatedSource, $namespace);
-    }
-
-    /**
-     * Process and reflect all the matching identifiers found inside a namespace node.
-     *
-     * @param Node\Stmt\Namespace_ $namespace
-     * @param IdentifierType $identifierType
-     * @param LocatedSource $locatedSource
-     * @return Reflection[]
-     */
-    private function reflectFromNamespace(
-        Node\Stmt\Namespace_ $namespace,
-        IdentifierType $identifierType,
-        LocatedSource $locatedSource
-    ) {
-        $reflections = [];
-        foreach ($namespace->stmts as $node) {
-            $reflection = $this->reflectNode($node, $locatedSource, $namespace);
-
-            if (null !== $reflection && $identifierType->isMatchingReflector($reflection)) {
-                $reflections[] = $reflection;
-            }
-        }
-        return $reflections;
-    }
-
-    /**
-     * Reflect identifiers from an AST. If a namespace is found, also load all the
-     * matching identifiers found in the namespace.
-     *
-     * @param Node[] $ast
-     * @param IdentifierType $identifierType
-     * @param LocatedSource $locatedSource
-     * @return \BetterReflection\Reflection\Reflection[]
-     */
-    private function reflectFromTree(array $ast, IdentifierType $identifierType, LocatedSource $locatedSource)
-    {
-        $reflections = [];
-        foreach ($ast as $node) {
-            if ($node instanceof Node\Stmt\Namespace_) {
-                $reflections = array_merge(
-                    $reflections,
-                    $this->reflectFromNamespace($node, $identifierType, $locatedSource)
-                );
-            } elseif ($node instanceof Node\Stmt\ClassLike) {
-                $reflection = $this->reflectNode($node, $locatedSource, null);
-                if ($identifierType->isMatchingReflector($reflection)) {
-                    $reflections[] = $reflection;
-                }
-            } elseif ($node instanceof Node\Stmt\Function_) {
-                $reflection = $this->reflectNode($node, $locatedSource, null);
-                if ($identifierType->isMatchingReflector($reflection)) {
-                    $reflections[] = $reflection;
-                }
-            }
-        }
-        return $reflections;
     }
 }
